+++ src/main/java/com/muzzy/dto/WalletDto.java
@@ -4,10 +4,12 @@ import lombok.Getter;
 import lombok.Setter;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Component
 public class WalletDto {
     private String publicKey;
-    private float balance;
+    private BigDecimal balance;
 }
