+++ src/main/java/com/muzzy/service/TransactionOutputService.java
@@ -3,6 +3,7 @@ package com.muzzy.service;
 import com.muzzy.domain.TransactionOutput;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.security.PublicKey;
 import java.util.Set;
 
@@ -13,5 +14,5 @@ import java.util.Set;
 @Service
 public interface TransactionOutputService extends CrudService<TransactionOutput,String> {
     Set<TransactionOutput> getTransctionByReciever(PublicKey publicKey);
-    float getBalance(PublicKey publicKey);
+    BigDecimal getBalance(PublicKey publicKey);
 }
